python/worker/requirements.txt
# Discord API
discord.py

# Database
tortoise-orm
asyncpg

# Redis
redis

# Data validation
pydantic

# Environment variables
python-dotenv

# Logging and utilities
python-json-logger
